2021/src/test/java/Day1Test.java
import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Nested;
import org.junit.jupiter.api.Test;

class Day1Test {

  @Nested
  class Teaser {

    Day puzzle = new Day1(
        "src/main/resources/day1-part1-teaser.txt");

    @Test
    void getAnswerPart1() {
      assertEquals(puzzle.getAnswerPart1(), 7);
    }

    @Test
    void getAnswerPart2() {
      assertEquals(puzzle.getAnswerPart2(), 5);
    }
  }

  @Nested
  class Real {

    Day puzzle = new Day1(
        "src/main/resources/day1.txt");

    @Test
    void getAnswerPart1() {
      assertEquals(puzzle.getAnswerPart1(), 1665);
    }

    @Test
    void getAnswerPart2() {
      assertEquals(puzzle.getAnswerPart2(), 1702);
    }
  }


}